refactor: decompose composite-entry.ts into focused modules
Split the 379-line god module into single-responsibility modules:

- arg-parser.ts: parseCompositeArgs, parseFlags, CompositeCommand type
- dispatcher-handler.ts: handleDispatcher, handleRedirect
- subskill-handler.ts: handleSubskill
- topic-handler.ts: handleTopics, handleTopic
- help-printer.ts: printCompositeHelp

composite-entry.ts is now an 82-line thin dispatcher that routes
parsed commands to the appropriate handler. Re-exports parseCompositeArgs
and CompositeCommand for backward compatibility with existing tests.
